Separate Cassandra-specific code from accessor
This makes the Accessor an abstract base class from which the Cassandra
and in-memory (test) Accessor inherit.
