+++ .travis.yml
@@ -17,7 +17,7 @@ env:
     - TOXENV=py27-coverage
   global:
     # As of 2016-04 Travis provides Cassandra 2.1 by default so we bundle our own
-    - CASSANDRA_VERSION=3.5
+    - CASSANDRA_VERSION=3.7
     - CASSANDRA_HOME="${TRAVIS_BUILD_DIR}/apache-cassandra-${CASSANDRA_VERSION}/"
 
 # As of 2016-04 Travis provides Cassandra 2.1 by default so we bundle our own
